Spring Boot: Fix health check in 2.2.x
This patch now only depends on the HealthIndicator class being present
on the classpath and delegates the responsibility of using the indicator
and registering it with a registry to Spring Boot itself.

Depending on a HealthIndicatorRegistry bean was a naive approach. In
Spring Boot 2.2.0, HealthIndicatorRegistry is deprecated in favor of
HealthContributorRegistry.