Adding in exponential failure backoff
Default uses a rate of 1.5. This was stolen from here
https://github.com/resilience4j/resilience4j/blob/03253134ac6f88954c8a65faa6dd7f92ed5e02f1/resilience4j-core/src/main/java/io/github/resilience4j/core/IntervalFunction.java#L20

But you can configure your own rate as well which allows for backoff to
be more or less aggressive as you see fit. This also works nicely with
the MaxRetriesFailureHandler since this is exponential retrying
potentially forever would make your retry times well days or potentially
scheduled out days, months, years in the future if they fail enough
times.