feat: fix lock-and-fetch for MySQL
The support was nearly finished in #371 but later commented out due to "some issue [..] and limited time".

I tried to find out what those issues were and discovered that the syntax of the select was wrong. In MySQL the `LIMIT` comes before the `FOR UPDATE SKIP LOCKED`.

Using that fix `Mysql8CompatibilityTest` passes, I have done no further testing yet.