feat: Add schedule-or-reschedule method to the SchedulerClient (#553)
## feat: Add rescheduleOrCreate method to the SchedulerClient
When invoking SchedulerClient.reschedule() with a non-existent
TaskInstance, an exception will be raised. Therefore, it's essential to
verify the existence of a scheduled TaskInstance prior to attempting a
reschedule. To streamline this process, we propose the introduction of a
scheduleOrCreate method. This method would automatically schedule a
TaskInstance if none exists, making it particularly beneficial for
dynamically recurring tasks that require initial setup.

We have given considerable thought to finding an appropriate name for
this method, but unfortunately, we haven't arrived at a satisfactory
conclusion. The verb "schedule" best describes its functionality;
however, it cannot be used as there is already a method with that name.
Since "rescheduleOrSchedule" sounds awkward, we have settled on
"rescheduleOrCreate" for now. We are open to considering other names for
this method.

## Fixes
[#539](https://github.com/kagkarlsson/db-scheduler/issues/539)


## Reminders
- [x] Added/ran automated tests
- [x] Update README and/or examples
- [x] Ran `mvn spotless:apply`

---------

Co-authored-by: Gustav Karlsson <kagkarlsson@gmail.com>