fix: Use SchedulerListerner instead of the deprecated StatsRegistry in the TaskResolver (#659)
## Description
This pull request is a follow-up to [PR
#504](https://github.com/kagkarlsson/db-scheduler/pull/504), where we
introduced the SchedulerListener interface as a replacement for the
StatsRegistry interface. While the migration had been initiated, it was
incomplete, resulting in the TaskResolver continuing to use
StatsRegistry, hence the SchedulerEventType.UNRESOLVED_TASK remained
unused.

Previously, reacting to the UNRESOLVED_TASK event required utilizing the
deprecated StatsRegistry. With this update, the StatsRegistry in the
TaskResolver is replaced with SchedulerListener, allowing support for
listening to UNRESOLVED_TASK events through both SchedulerListener and
StatsRegistry (via the StatsRegistryAdapter). This change not only
enables the intended functionality but also maintains backward
compatibility.


## Fixes
Fixes #658 


## Reminders
- [X] Added/ran automated tests
- [X] Update README and/or examples
- [X] Ran `mvn spotless:apply`

---
cc @kagkarlsson

---------

Co-authored-by: Gustav Karlsson <kagkarlsson@gmail.com>