fix: Daily Schedule should add day on LocalDate before setting the LocalTi… (#663)
When using a daily schedule with a time like 01:30 in the Europe/London
timezone, the DST transition on March 30, 2025 causes an issue. That day
the schedule triggers at 02:30, since 01:30 does not exist on that day
due to the clock jumping forward. This is a reasonable behavior
(inherited ZonedDateTime#of method behavior). However, the bug arises in
the next scheduled iteration, which also returns 02:30 instead of
reverting to the correct 01:30.

## Fixes
Fixes #662 

## Reminders
- [x] Added/ran automated tests
- [x] Update README and/or examples
- [x] Ran `mvn spotless:apply`

---
Co-authored-by: Gustav Karlsson <kagkarlsson@gmail.com>