feat: Expose Clock from builder and implement in Spring Boot AutoConfiguration (#720)
## Brief, plain english overview of your changes here
- Added clock(Clock) method to SchedulerBuilder so users can provide
their own clock for Scheduler
- In order not to have to think about Waiter caching old Clock I moved
Waiter creation to build method, so it always uses latest Clock from
builder field
- Had to build Waiter manually in TestHelper for ManualScheduler.

Note: This changes current behavior, but I think it's a bug. 
Currently Waiter in ManualScheduler is using SystemClock instead of
newly set clock, I changed that because Waiter is now created in the end
so it will always use up-to date Clock instance, in case this is desired
behavior for some reason, let me know and I will fix it.


## Reminders
- [x] Added/ran automated tests
- [ ] Update README and/or examples
- [x] Ran `mvn spotless:apply`

---
cc @kagkarlsson

---------

Co-authored-by: Gustav Karlsson <kagkarlsson@gmail.com>