feat: allow arbitrary tasks chaining (#728)
`ExecutionOperations.removeAndScheduleNew` accepts
`SchedulableInstance<T>`, when in fact there is no reason for it to not
be able to schedule arbitrary tasks.




This caused me problems when I was trying to hack in the DLQ support, my
idea was to do something in this fashion:

```
FailureHandler.MaxRetriesFailureHandler<T>(
    maxRetries,
    FailureHandler.ExponentialBackoffFailureHandler(duration),
    { executionComplete, executionOperations ->
        moveToDlq(executionComplete, executionOperations)
    }
)

fun <T> moveToDlq(executionComplete: ExecutionComplete, executionOperations: ExecutionOperations<T>) {
    val ti = executionComplete.execution.taskInstance
    val schedulableInstance = DLQ.instance("")
        .data(
            DlqJob(
                ti.id,
                ti.taskName,
                ti.data
            )
        )
        .scheduledTo(Instant.MAX)
    @Suppress("UNCHECKED_CAST")
    executionOperations
        .removeAndScheduleNew(schedulableInstance as SchedulableInstance<T?>?)
```

My alternate idea was copying the task to another table. The issue with
it was, `ExecutionOperations` create their own transactions.

---------

Co-authored-by: Gustav Karlsson <kagkarlsson@gmail.com>